Add CI to the velero-plugin-for-vsm
Use github action for the velero-plugin-for-vsm plugin.

This project is within migtools where Github actions are allowed
and as such it's easy to run go test is used.

We do not want to run make test as this project does not have
all the Makefile deps and folders for building it.

Signed-off-by: Michal Pryc <mpryc@redhat.com>